Develop

+++ src/main/java/com/kakaoasset/portfolio/entity/auth/OauthToken.java
@@ -7,7 +7,7 @@ public class OauthToken {
     private String access_token;
     private String token_type;
     private String refresh_token;
-    private int expires_in;
+    private Long expires_in;
     private String scope;
-    private int refresh_token_expires_in;
+    private Long refresh_token_expires_in;
 }
